updated elections
elections should only be held when follower does not know who leader is. Also don't want to start leader heart-beats to followers. If a follower is still a part of the network, it okay that follower might be unaware of leader. Follower can ask who is leader when needing to call leader (using leader-port/leader-host)

Need a better way to track who is still on the network so that leader can ask coordination server for list of all connections